Move createSignedOrder back to actions.
Invoke getUserOrders & getAllOrders on successful order submit (action).
